Remove python from packages.yaml
Remove python from packages.yaml. The current modulefile for python has
two paths set to $PATH. This confuses Spack and leads to it picking up
the wrong executable (mpi4py) instead of the python exe.

For the python installation, ensure that we are doing dynamic linking
with python. Users will most likely need to specify ^python+shared for
anything that depends on python (direct deps only).